requirements.txt
#python 3.8.1

aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
easyocr==1.7.1
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
openai==0.27.8
opencv-python-headless==4.5.4.60
outcome==1.3.0.post0
packaging==23.2
Pillow==9.5.0
plyer==2.1.0
pyclipper==1.3.0.post5
pycparser==2.21
pypdfium2==4.26.0
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
selenium==4.16.0
shapely==2.0.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
tifffile==2023.7.10
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
